pyproject.toml
[project]
name = "vectordb-etl"
version = "0.1.0"
description = "VectorDB ETL Pipeline for HTML documents to Milvus"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
keywords = ["vectordb", "milvus", "etl", "embeddings", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.38",
    "langchain-core>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "pymilvus[model]>=2.4.0",
    "sentence-transformers>=2.3.1",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.8.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
vectordb-etl = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["modules"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
